+++ test/libsolidity/syntaxTests/immutable/conditionally_initialized.sol
@@ -5,5 +5,3 @@ contract C {
             x = 1;
     }
 }
-// ----
-// TypeError 4599: (86-87): Cannot write to immutable here: Immutable variables cannot be initialized inside an if statement.