+++ test/libsolidity/syntaxTests/immutable/ctor_initialization_indirect_reading.sol
@@ -6,5 +6,3 @@ contract C {
 
     function f() public view returns (uint) { return 3 + x; }
 }
-// ----
-// TypeError 7733: (136-137): Immutable variables cannot be read before they are initialized.