+++ test/libsolidity/syntaxTests/immutable/decrement.sol
@@ -4,5 +4,3 @@ contract C {
         x--;
     }
 }
-// ----
-// TypeError 3969: (63-64): Immutable variables must be initialized using an assignment.