+++ test/libsolidity/syntaxTests/immutable/inheritance_ctor_argument.sol
@@ -10,5 +10,3 @@ contract C is B {
     uint immutable y;
     constructor() B(y = 3) { }
 }
-// ----
-// TypeError 1581: (148-149): Cannot write to immutable here: Immutable variables can only be initialized inline or assigned directly in the constructor.