+++ test/libsolidity/syntaxTests/immutable/inheritance_ctor_inherit_specifier_argument_init.sol
@@ -9,5 +9,3 @@ contract B {
 contract C is B(C.y = 3) {
     uint immutable y;
 }
-// ----
-// TypeError 1581: (104-107): Cannot write to immutable here: Immutable variables can only be initialized inline or assigned directly in the constructor.