+++ test/libsolidity/syntaxTests/immutable/init_in_constructor_modifier_param.sol
@@ -4,4 +4,3 @@ contract D {
     constructor() m(t = 2) {}
 }
 // ----
-// TypeError 1581: (83-84): Cannot write to immutable here: Immutable variables can only be initialized inline or assigned directly in the constructor.