+++ test/libsolidity/syntaxTests/immutable/read_in_function_before_init.sol
@@ -8,5 +8,3 @@ contract B is A {
 	uint immutable x = 2;
 	function f() override pure internal returns (uint) { return x; }
 }
-// ----
-// TypeError 7733: (223-224): Immutable variables cannot be read before they are initialized.